Modified script names for test.
